Security hardening: ~119 memory-safety & robustness corrections
Guest->host out-of-bounds / NULL-deref fixes across device, loader, PROM,
disk and network paths; an author-ethos course-correction (warn loudly and
continue rather than silently mask or crash); and additive PPC extended-BAT
(745x) support. See CHANGELOG.md and REVIEW_FINDINGS.md for the per-fix
breakdown.

Validated: gcc -O3 -Wall -Wextra and clang-21 both 0 errors / 0 warnings;
regression battery across 5 CPU arches (boot + device-attach), a full
OpenBSD/pmax multiuser rig (root, NAT, disk, daemons, reboot-persist), and
OpenBSD/luna88k (M88K) + NetBSD/landisk (SH4) interactive shells.

GXemul is an educational/experimental emulator, not a security boundary;
these are robustness/memory-safety hardening fixes, not a sandbox-escape
disclosure. Not affiliated with or endorsed by Anders Gavare.

+++ src/cpus/cpu_dyntrans.c
@@ -1698,7 +1698,7 @@ void DYNTRANS_UPDATE_TRANSLATION_TABLE(struct cpu *cpu, uint64_t vaddr_page,
 		index = DYNTRANS_ADDR_TO_PAGENR(vaddr_page);
 		cpu->cd.DYNTRANS_ARCH.phys_page[index] = NULL;
 #ifdef DYNTRANS_ARM
-		cpu->cd.DYNTRANS_ARCH.is_userpage[index>>5] &= ~(1<<(index&31));
+		cpu->cd.DYNTRANS_ARCH.is_userpage[index>>5] &= ~((uint32_t)1<<(index&31));
 		if (useraccess)
 			cpu->cd.DYNTRANS_ARCH.is_userpage[index >> 5]
 			    |= 1 << (index & 31);
@@ -1833,7 +1833,7 @@ cpu->cd.DYNTRANS_ARCH.vph_tlb_entry[r].valid);
 		x /= addr_per_translation_range;
 
 		cpu->cd.DYNTRANS_ARCH.cur_physpage->
-		    translations_bitmap |= (1 << x);
+		    translations_bitmap |= ((uint32_t)1 << x);
 	}
